Add config schema updater & docs (#2056)
* Add config schema updater & docs

* Updates following review; fix example

* Add instructions on how to check for prior changes to config item

* Lower case `name`; Actually use `name`

* Not allowed to use `Field` with `_name` attr; mass schema file rename

* Simplify schema name setting
