+++ REUSE.toml
@@ -21,7 +21,7 @@ path = [
     "envs/*.pin.txt",
     "envs/environment.yaml",
     "config/config.default.yaml",
-    "config/schema.json"
+    "config/schema.default.json"
 ]
 SPDX-FileCopyrightText = "The PyPSA-Eur Authors"
 SPDX-License-Identifier = "CC0-1.0"
