+++ config/config.default.yaml
@@ -1,6 +1,6 @@
 %YAML 1.1
 ---
-# yaml-language-server: $schema=./schema.json
+# yaml-language-server: $schema=./schema.default.json
 
 # docs in https://pypsa-eur.readthedocs.io/en/latest/configuration.html#version
 version: v2025.07.0
