+++ scripts/lib/validation/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: Contributors to PyPSA-Eur <https://github.com/pypsa/pypsa-eur>
+#
+# SPDX-License-Identifier: MIT
+
+import scripts.lib.validation.config_updates
